move helpers to test-helpers file
