+++ tests/test-helpers.js
@@ -3,6 +3,7 @@ import { dirname, join } from 'path';
 import { packageUp } from 'package-up';
 import fixturify from 'fixturify';
 import stripAnsi from 'strip-ansi';
+import tmp from 'tmp';
 
 export async function getCliPath(version) {
   const path = dirname(require.resolve(version));
@@ -85,4 +86,39 @@ export async function testWithTestem(cwd, expect, testemPort) {
 
   expect(result.exitCode, result.output).to.equal(0);
   console.log(result.stdout);
+}
+
+export const testVersions = [
+  // ['ember-cli-3.28'],
+  // ['ember-cli-4.12'],
+  // ['ember-cli-4.4'],
+  // ['ember-cli-4.8'],
+  // // test helpers seems to be broken for most ember versions 😭
+  // ['ember-cli-5.4', ['@ember/test-helpers@latest']],
+  // ['ember-cli-5.8', ['@ember/test-helpers@latest']],
+  ['ember-cli-5.12', ['@ember/test-helpers@latest']],
+  ['ember-cli-latest'],
+];
+
+let port = 7357;
+export function getPort() {
+  return port++;
+}
+
+export async function executeTest(
+  expect,
+  version,
+  packages,
+  cliOptions,
+  testemPort,
+) {
+  let tmpobj = tmp.dirSync({ unsafeCleanup: true });
+  const cwd = join(tmpobj.name, 'test-app');
+  const cliPath = await getCliPath(version);
+
+  await generateEmberApp(tmpobj.name, version, packages, cliPath, cliOptions);
+  await testEmber(cwd, expect, testemPort);
+  await runCodemod(cwd);
+  await testEmber(cwd, expect, testemPort);
+  await testWithTestem(cwd, expect, testemPort);
 }
